updated test cases

+++ servo-core/src/main/java/com/netflix/servo/util/Counters.java
@@ -66,7 +66,7 @@ public final class Counters {
      * @param name   name of the counter to increment
      * @param delta  the amount to increment the counter by
      */
-    public static void increment(String name, int delta) {
+    public static void increment(String name, long delta) {
         increment(new MetricConfig(name), delta);
     }
 
@@ -87,7 +87,7 @@ public final class Counters {
      * @param tags   tags to associate with the counter
      * @param delta  the amount to increment the counter by
      */
-    public static void increment(String name, TagList tags, int delta) {
+    public static void increment(String name, TagList tags, long delta) {
         increment(new MetricConfig(name, tags), delta);
     }
 
@@ -97,7 +97,7 @@ public final class Counters {
      * @param config  config of the counter to increment
      * @param delta   the amount to increment the counter by
      */
-    public static void increment(MetricConfig config, int delta) {
+    public static void increment(MetricConfig config, long delta) {
         TagList cxtTags = TaggingContext.getTags();
         if (cxtTags != null) {
             String name = config.getName();
